Modal panel service refactor
